Move interfaces to models.ts
